Kanvi/Capture last prefetch (#345)
Capture the prefetchdataset op in certain cases to allow data pipelining
