+++ ngraph_bridge/ngraph_utils.cc
@@ -473,6 +473,41 @@ std::string GraphFilenamePrefix(std::string kind, int idx, int sub_idx) {
   return ss.str();
 }
 
+void DumpGraphs(const GraphOptimizationPassOptions& options, int idx,
+                std::string filename_prefix, std::string title) {
+  // If we have a "main" graph, dump that.
+  if (options.graph != nullptr) {
+    auto dot_filename = DotFilename(filename_prefix, idx);
+    auto pbtxt_filename = PbtxtFilename(filename_prefix, idx);
+    NGRAPH_VLOG(0) << "Dumping main graph to " << dot_filename;
+    NGRAPH_VLOG(0) << "Dumping main graph to " << pbtxt_filename;
+
+    GraphToDotFile(options.graph->get(), dot_filename, title);
+    GraphToPbTextFile(options.graph->get(), pbtxt_filename);
+  }
+
+  // If we have partition graphs (we shouldn't), dump those.
+  if (options.partition_graphs != nullptr) {
+    int sub_idx = 0;
+
+    for (auto& kv : *options.partition_graphs) {
+      auto dot_filename = DotFilename(filename_prefix, idx, sub_idx);
+      auto pbtxt_filename = PbtxtFilename(filename_prefix, idx, sub_idx);
+      NGRAPH_VLOG(0) << "Dumping subgraph " << sub_idx << " to "
+                     << dot_filename;
+      NGRAPH_VLOG(0) << "Dumping subgraph " << sub_idx << " to "
+                     << pbtxt_filename;
+
+      Graph* pg = kv.second.get();
+
+      GraphToDotFile(pg, dot_filename, title);
+      GraphToPbTextFile(pg, pbtxt_filename);
+
+      sub_idx++;
+    }
+  }
+}
+
 bool DumpAllGraphs() { return std::getenv("NGRAPH_TF_DUMP_GRAPHS") != nullptr; }
 
 bool DumpPrecaptureGraphs() {
