+++ test/test_utilities.cpp
@@ -313,6 +313,16 @@ Status LoadGraph(const string& graph_file_name,
   return (*session)->Create(graph_def);
 }
 
+Status LoadGraphFromPbTxt(const string& pb_file, Graph* input_graph) {
+  // Read the graph
+  tensorflow::GraphDef graph_def;
+  TF_RETURN_IF_ERROR(ReadTextProto(Env::Default(), pb_file, &graph_def));
+  GraphConstructorOptions opts;
+  opts.allow_internal_ops = true;
+  auto status = ConvertGraphDefToGraph(opts, graph_def, input_graph);
+  return status;
+}
+
 template <>
 void AssignInputValues(Tensor& A, int8 x) {
   auto A_flat = A.flat<int8>();
